% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_angle_difference}
\alias{add_angle_difference}
\title{Adds new colum angle_diff_axis where it calculates angle difference between rows}
\usage{
add_angle_difference(df_log, rotation, name)
}
\arguments{
\item{df_log}{data frame log that gets modified}

\item{name}{what should be appended to "angle_diff_"}
}
\description{
Adds new colum angle_diff_axis where it calculates angle difference between rows
}
