% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_angle_differences}
\alias{add_angle_differences}
\title{Adds extra diff column to each column that has "rotation" in its name}
\usage{
add_angle_differences(variables)
}
\arguments{
\item{obj}{}
}
\description{
Adds extra diff column to each column that has "rotation" in its name
}
