% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-loading.R
\name{add_area_boundaries}
\alias{add_area_boundaries}
\title{Validates and adds area boudaries}
\usage{
add_area_boundaries(obj, ls)
}
\arguments{
\item{obj}{navr object}

\item{ls}{list of list(x = c(0, 100), y=c(0,100)) type. Needs to have
x and y boundaries for the graphs to work properly}
}
\value{
modified object
}
\description{
Validates and adds area boudaries
}
