% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_columns_navr}
\alias{add_columns_navr}
\title{Adds all necessary columns to the navr object}
\usage{
add_columns_navr(obj)
}
\arguments{
\item{obj}{object of type navr}
}
\value{
returns navr object with modified columns
}
\description{
runs add_time_columns, add_distances, add_angle_differences and add_speeds
}
