% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_distances}
\alias{add_distances}
\title{Adds distance and cumulative distances (distance_total) columns}
\usage{
add_distances(obj, ...)
}
\arguments{
\item{obj}{valid navr object}

\item{...}{}
}
\value{
navr object
}
\description{
Adds distance and cumulative distances (distance_total) columns
}
