% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_navr.R
\name{add_font_awesome}
\alias{add_font_awesome}
\title{Easy Icons for \code{navr}}
\usage{
add_font_awesome(nav = NULL)
}
\value{
\code{navr} htmlwidget with Font-Awesome dependencies attached.
}
\description{
Add \href{http://fontawesome.io/}{Font-Awesome} for styling
  our \code{navr} with beautiful, professional icons.  Please note
  that you'll already have these icons if using Shiny.
}

