% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-modifying.R
\name{add_goal_order}
\alias{add_goal_order}
\title{Adds goal order to the object}
\usage{
add_goal_order(obj, order, ...)
}
\arguments{
\item{order}{vector with goal order. Indexed from 1 to n:rows of goal positions}

\item{...}{aditional arguments}
}
\value{
modified object
}
\description{
Adds goal order to the object
}
