% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-modifying.R
\name{add_goal_positions}
\alias{add_goal_positions}
\title{Adds goal positions to the object}
\usage{
add_goal_positions(obj, df, ...)
}
\arguments{
\item{df}{data frame with goal positions with appropriate column naming}

\item{...}{aditional arguments}
}
\value{
modified object
}
\description{
Adds goal positions to the object
}
