% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_navr.R
\name{add_hover}
\alias{add_hover}
\title{Hover Effects for \code{navr}}
\usage{
add_hover(nav = NULL, effect = NULL)
}
\arguments{
\item{nav}{\code{navr} htmlwidget to which we will apply hover effects}

\item{effect}{\code{string} hover effect which we would like to apply.  For options,
         see \href{http://ianlunn.github.io/Hover/}{Hover Effects}.}
}
\value{
\code{navr} htmlwidget with hover effects applied and dependencies attached.
}
\description{
Add \href{http://ianlunn.github.io/Hover/}{Hover Effects} to
 our \code{navr}.  There are lots of hover effects.
}

