% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_speeds}
\alias{add_speeds}
\title{Adds speed column}
\usage{
add_speeds(obj, ...)
}
\arguments{
\item{obj}{valid navr object}

\item{...}{}
}
\value{
navr object
}
\description{
Adds speed column
}
