% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_time_columns}
\alias{add_time_columns}
\title{Add time_since_start and time_diff columns}
\usage{
add_time_columns(obj, ...)
}
\arguments{
\item{obj}{Navr Object}
}
\description{
Add time_since_start and time_diff columns
}
