% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{add_times_since_start}
\alias{add_times_since_start}
\title{Adds tiems since start column to the navr object}
\usage{
add_times_since_start(obj, ...)
}
\arguments{
\item{obj}{object of appropriate class}

\item{...}{}
}
\value{
object with its
}
\description{
Adds tiems since start column to the navr object
}
