% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{angle_diff}
\alias{angle_diff}
\title{Calculates angular difference between passed angle matrices}
\usage{
angle_diff(angle1, angle2)
}
\arguments{
\item{angle1}{vector of angles (in degrees)}

\item{angle2}{}
}
\description{
Calculates angular difference between passed angle matrices
}
