% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{angle_to_180}
\alias{angle_to_180}
\title{Converts angle to -180 to 180 difference}
\usage{
angle_to_180(angle)
}
\arguments{
\item{angle}{}
}
\value{
number between -180 to 180
}
\description{
Converts angle to -180 to 180 difference
}
