% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{angle_to_360}
\alias{angle_to_360}
\title{converts positive and negative angles to 0-360}
\usage{
angle_to_360(angle)
}
\arguments{
\item{angle}{}
}
\description{
asumes it is not below -360
390 is converted to 30, -40 to 320 etc
}
