% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{angle_to_radian}
\alias{angle_to_radian}
\title{Converts angle to radian}
\usage{
angle_to_radian(angle)
}
\arguments{
\item{angle}{}
}
\value{
radians
}
\description{
Converts angle to radian
}
