% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{animate_path}
\alias{animate_path}
\title{Plot built with plot_path function}
\usage{
animate_path(plt, ...)
}
\arguments{
\item{plt}{Needs to have timestamp x value}

\item{...}{}
}
\description{
Plot built with plot_path function
}
