% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anslag_avrund.R
\name{anslag_avrund}
\alias{anslag_avrund}
\title{anslag_avrund}
\usage{
anslag_avrund(x, i_mill_kr = F)
}
\arguments{
\item{x}{verdi som skal avrundes.}

\item{i_mill_kr}{boolean TRUE eller FALSE i mill.kr eller i hele kroner. Default er FALSE}
}
\description{
Avrunder etter avrundingsregler til budsjettanslag til Budsjettgruppa.
}
\details{
Avrunder verdi etter avrundingsreglene til hvordan budsjettanslag skal leveres til Budsjettgruppa.
*Poster på under 2 mill. kroner blir avrundet til nærmeste 0,1 mill. kroner.
*Poster i intervallet 2–199 mill. kroner blir avrundet til nærmeste 1 mill. kroner.
*Poster i intervallet 200–999 mill. kroner blir avrundet til nærmeste 5 mill. kroner.
*Poster på 1000 mill. kroner eller mer blir avrundet til nærmeste 10 mill. kroner
}
\examples{
anslag_avrund(x = 204.1, i_mill_kr = T)
returns 205
}
