% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{AreaObject}
\alias{AreaObject}
\title{Definition of an area object used in designating area presence}
\usage{
AreaObject(name, type = "rectangle", points = c())
}
\arguments{
\item{name}{character. Name of the area}

\item{type}{character. defining what type of area. Currently supported is: "rectangle"}

\item{points}{numeric matrix (npoints, 2) with each row representing a single point and
columns representing X and Y positions}
}
\description{
Definition of an area object used in designating area presence
}
\examples{
\dontrun{
AreaObject("main square", "rectangle", matrix(c(0, 0, 1, 0, 1, 1, 0, 1), ncol=2, byrow=TRUE))
}
}
