% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Budsjett.R
\name{Budsjett}
\alias{Budsjett}
\title{Budsjettobjektet.}
\description{
Hovedobjektet til budsjettdokumentet.
Budsjett-objektet har metoder som lager tabeller, og navn til tabeller og tekst.
Et budsjett-objekt for hver post.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Budsjett$new()}}
\item \href{#method-lagRegnskapTabell}{\code{Budsjett$lagRegnskapTabell()}}
\item \href{#method-lagRegnskapTabell2}{\code{Budsjett$lagRegnskapTabell2()}}
\item \href{#method-lagMottakerTabell}{\code{Budsjett$lagMottakerTabell()}}
\item \href{#method-giRegnskapstall}{\code{Budsjett$giRegnskapstall()}}
\item \href{#method-giRegnskapstallIfjor}{\code{Budsjett$giRegnskapstallIfjor()}}
\item \href{#method-lagTabellMndUtviklingRegnskap}{\code{Budsjett$lagTabellMndUtviklingRegnskap()}}
\item \href{#method-lagTabellMndUtviklingMottakere}{\code{Budsjett$lagTabellMndUtviklingMottakere()}}
\item \href{#method-setKapittelpost}{\code{Budsjett$setKapittelpost()}}
\item \href{#method-giKapittelpost}{\code{Budsjett$giKapittelpost()}}
\item \href{#method-leggTilNyttAnslag}{\code{Budsjett$leggTilNyttAnslag()}}
\item \href{#method-giPeriode}{\code{Budsjett$giPeriode()}}
\item \href{#method-giAnslag}{\code{Budsjett$giAnslag()}}
\item \href{#method-leggTilHistoriskAnslag}{\code{Budsjett$leggTilHistoriskAnslag()}}
\item \href{#method-giHistoriskeAnslag}{\code{Budsjett$giHistoriskeAnslag()}}
\item \href{#method-print}{\code{Budsjett$print()}}
\item \href{#method-clone}{\code{Budsjett$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$new(
  name,
  name_nytt_anslag,
  periode,
  dfRegnskap,
  pris_gjeldende,
  dfMottakere,
  PRIS_VEKST,
  nyeAnslag = NULL,
  historiskeAnslag = NULL,
  kapittelpost = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Det lange og spesifiserte navnet på anslaget.}

\item{\code{name_nytt_anslag}}{kortnavnet eks. "oktober-anslaget"}

\item{\code{kapittelpost}}{kapittel og post, format som 2620.70}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagRegnskapTabell"></a>}}
\if{latex}{\out{\hypertarget{method-lagRegnskapTabell}{}}}
\subsection{Method \code{lagRegnskapTabell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$lagRegnskapTabell()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagRegnskapTabell2"></a>}}
\if{latex}{\out{\hypertarget{method-lagRegnskapTabell2}{}}}
\subsection{Method \code{lagRegnskapTabell2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$lagRegnskapTabell2(
  anslag1,
  anslag2,
  printversjon = FALSE,
  celing_date = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagMottakerTabell"></a>}}
\if{latex}{\out{\hypertarget{method-lagMottakerTabell}{}}}
\subsection{Method \code{lagMottakerTabell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$lagMottakerTabell()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giRegnskapstall"></a>}}
\if{latex}{\out{\hypertarget{method-giRegnskapstall}{}}}
\subsection{Method \code{giRegnskapstall()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giRegnskapstall(ar_til_bake)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giRegnskapstallIfjor"></a>}}
\if{latex}{\out{\hypertarget{method-giRegnskapstallIfjor}{}}}
\subsection{Method \code{giRegnskapstallIfjor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giRegnskapstallIfjor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagTabellMndUtviklingRegnskap"></a>}}
\if{latex}{\out{\hypertarget{method-lagTabellMndUtviklingRegnskap}{}}}
\subsection{Method \code{lagTabellMndUtviklingRegnskap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$lagTabellMndUtviklingRegnskap(
  tiltak_kost_ar1 = NULL,
  tiltak_kost_ar2 = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagTabellMndUtviklingMottakere"></a>}}
\if{latex}{\out{\hypertarget{method-lagTabellMndUtviklingMottakere}{}}}
\subsection{Method \code{lagTabellMndUtviklingMottakere()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$lagTabellMndUtviklingMottakere()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setKapittelpost"></a>}}
\if{latex}{\out{\hypertarget{method-setKapittelpost}{}}}
\subsection{Method \code{setKapittelpost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$setKapittelpost(kapittelpost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kapittelpost}}{kapittel og post, format som 2620.70}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giKapittelpost"></a>}}
\if{latex}{\out{\hypertarget{method-giKapittelpost}{}}}
\subsection{Method \code{giKapittelpost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giKapittelpost(kapittelpost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kapittelpost}}{kapittel og post, format som 2620.70}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-leggTilNyttAnslag"></a>}}
\if{latex}{\out{\hypertarget{method-leggTilNyttAnslag}{}}}
\subsection{Method \code{leggTilNyttAnslag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$leggTilNyttAnslag(anslag, rekkefolge)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giPeriode"></a>}}
\if{latex}{\out{\hypertarget{method-giPeriode}{}}}
\subsection{Method \code{giPeriode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giPeriode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giAnslag"></a>}}
\if{latex}{\out{\hypertarget{method-giAnslag}{}}}
\subsection{Method \code{giAnslag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giAnslag()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-leggTilHistoriskAnslag"></a>}}
\if{latex}{\out{\hypertarget{method-leggTilHistoriskAnslag}{}}}
\subsection{Method \code{leggTilHistoriskAnslag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$leggTilHistoriskAnslag(anslag, rekkefolge)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giHistoriskeAnslag"></a>}}
\if{latex}{\out{\hypertarget{method-giHistoriskeAnslag}{}}}
\subsection{Method \code{giHistoriskeAnslag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$giHistoriskeAnslag()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Budsjett$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
