% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{calculate_angle_differences}
\alias{calculate_angle_differences}
\title{Adds new colum angle_diff_axis where it calculates angle difference between rows}
\usage{
calculate_angle_differences(rotations)
}
\arguments{
\item{rotations}{vector of angles in 360}
}
\description{
Adds new colum angle_diff_axis where it calculates angle difference between rows
}
