% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{calculate_areas_time}
\alias{calculate_areas_time}
\title{Calculates time spent in each of encoded areas}
\usage{
calculate_areas_time(obj)
}
\arguments{
\item{obj}{preprocessed navr object with added areas (see \code{\link{add_areas}})}
}
\value{
dataframe with results
}
\description{
Calculates time spent in each of encoded areas
}
