% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-distance-calculations.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculates distances froma x and y vectors}
\usage{
calculate_distances(mat_xy, first_value = NULL)
}
\arguments{
\item{first_value}{value to be appended to the first position (typically NA, or 0), if NULL, nothing is appended}

\item{positions_x}{vector of consecutive x positions}

\item{positions_y}{vector of consecutive y positions}
}
\value{
vector of distances of length x-1 or x if first value is defined
}
\description{
Calculates distances froma x and y vectors
}
