% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-speed-calculations.R
\name{calculate_speeds}
\alias{calculate_speeds}
\title{Calcualtes vector of doubles with speeds}
\usage{
calculate_speeds(distances, timestamps)
}
\arguments{
\item{distances}{distance measures in any unit}

\item{timestamps}{timestamps of the distance measurements}
}
\value{
vector of immediate speeds in given unit per second
}
\description{
Calcualtes vector of doubles with speeds
}
