% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-time-calculations.R
\name{calculate_times_diffs}
\alias{calculate_times_diffs}
\title{Calculates time differences between timestamps}
\usage{
calculate_times_diffs(timestamps, first_value = NULL)
}
\arguments{
\item{timestamps}{vector of timestamps in seconds}

\item{first_value}{what to insert in the first position (default NULL)}
}
\value{
vector of size of timestamps-1 or timestamps in case first value is set
}
\description{
Calculates time differences between timestamps
}
