% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-time-calculations.R
\name{calculate_times_since_start}
\alias{calculate_times_since_start}
\title{Calculates times since start}
\usage{
calculate_times_since_start(timestamps)
}
\arguments{
\item{timestamps}{timestamps in seconds}
}
\value{
vector of size of timestamps (startign with 0)
}
\description{
Calculates times since start
}
