% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-visualising.R
\name{create_direction_line}
\alias{create_direction_line}
\title{Creates arrow pointing from a certain position and angle}
\usage{
create_direction_line(position, angle, len)
}
\arguments{
\item{len}{}
}
\value{

}
\description{
Creates arrow pointing from a certain position and angle
}
