% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-complex-plots.R
\name{create_plot}
\alias{create_plot}
\title{Creates empty plot with invisible theme to clearly plot paths and points}
\usage{
create_plot()
}
\description{
Creates empty plot with invisible theme to clearly plot paths and points
}
