% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{create_void_plot}
\alias{create_void_plot}
\title{Creates empty plot with invisible theme to clearly plot paths and points}
\usage{
create_void_plot()
}
\description{
Creates empty plot with invisible theme to clearly plot paths and points
}
