% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-helpers.R
\name{euclid_distance}
\alias{euclid_distance}
\title{Euclidian distance between two points}
\usage{
euclid_distance(point1, point2)
}
\arguments{
\item{point1}{2D vector of X and Y position}

\item{point2}{2D vector of X and Y position}
}
\value{
numeric or NA if not valid
}
\description{
Calculates euclidian distance between two points.
}
\examples{
> print(euclid_distance(c(0,0), c(1,1)))
[1] 1.414214
}
