% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-distance-calculations.R
\name{euclid_distance_between_rows}
\alias{euclid_distance_between_rows}
\title{Eucilidan distance covered between each row}
\usage{
euclid_distance_between_rows(points)
}
\arguments{
\item{points}{m x 2 matrix or data.frame with first column X and second Y coordinates}
}
\value{
vector of distances of length m - 1
}
\description{
Calculates distance between points in each row
}
\examples{

}
