% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{geom_navr_area}
\alias{geom_navr_area}
\title{Returns area ploygon to be plotted}
\usage{
geom_navr_area(area, fill = NA, color = "red", size = 1.25, ...)
}
\arguments{
\item{area}{AreaObject}

\item{...}{other ggplot arguments for geom_polygon}
}
\value{
ggplot geom_polygon object
}
\description{
Returns area ploygon to be plotted
}
