% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_background}
\alias{geom_navr_background}
\title{GGplot geom to add custom PNG background to the plot}
\usage{
geom_navr_background(image_path, xlim = NULL, ylim = NULL)
}
\arguments{
\item{image_path}{path the png image}

\item{xlim}{background poistions as vector of length 2, c(-100,400)}

\item{ylim}{background poistions as vector of length 2, c(-100,400)}
}
\value{

}
\description{
GGplot geom to add custom PNG background to the plot
}
