% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_circle}
\alias{geom_navr_circle}
\title{Creates geom of a circle to be inserted into the graph}
\usage{
geom_navr_circle(center, radius, precision = 100, ...)
}
\arguments{
\item{center}{circle center as a numeric(2) \emph{default}: c(0,0)}

\item{radius}{circle radius}

\item{precision}{how many points will make the circle}

\item{...}{}
}
\description{
Creates geom of a circle to be inserted into the graph
}
