% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_direction}
\alias{geom_navr_direction}
\title{Adds arrow pointing from a point in a specified angle}
\usage{
geom_navr_direction(position, angle, length = 1, ...)
}
\arguments{
\item{position}{numeric(2) of X and Y position}

\item{angle}{numeric angle in degrees (0-360)}

\item{length}{length od the arrow to be drawm}

\item{...}{additional ggplot values}
}
\value{
ggplot element
}
\description{
Adds arrow pointing from a point in a specified angle
}
\examples{
geom_navr_direction(c(0,0), 180, 5, color = "red")
}
