% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_heatmap}
\alias{geom_navr_heatmap}
\title{geom to add stat_density2d position heatmap}
\usage{
geom_navr_heatmap(obj, bins = 25, ...)
}
\arguments{
\item{obj}{navr object}

\item{bins}{number of bins in each direction (n parameter in statn_density)}

\item{...}{}
}
\description{
geom to add stat_density2d position heatmap
}
