% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_limits}
\alias{geom_navr_limits}
\title{Adds limits to the plot from the area_boundaries list field}
\usage{
geom_navr_limits(obj)
}
\arguments{
\item{obj}{Navr object with area_boundaries field}
}
\value{
list with xlim and ylim based on object area boundaries
}
\description{
for better control, just use regular xlim and ylim functions, this is just a shorthand
}
