% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path}
\alias{geom_navr_path}
\title{Geom of plotting navr path}
\usage{
geom_navr_path(obj, ...)
}
\arguments{
\item{...}{}

\item{x}{vector of positions}

\item{y}{vector of y positions}
}
\description{
Geom of plotting navr path
}
