% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_events}
\alias{geom_navr_path_events}
\title{Plots positions of given events on a path}
\usage{
geom_navr_path_events(obj, event_times, size = 2, shape = 18,
  color = "blue", ...)
}
\arguments{
\item{obj}{navr object}

\item{event_times}{matrix of event times. It can have one or two dimensions,
if single events or events with a duration are to be plotted
\item{one dimension}{Each number represents a time in which an event occured.
See \code{\link{geom_navr_path_points}}}
\item{two dimentsions}{Each column of the matrix marks the begining and an end
of the event. First row is the beginings and second is the ends. Path between
the beginning and end is then colored. See \code{\link{geom_navr_path_segments}}}
Times need to be on the same scale as the timestamps in \code{obj$data$timestamp}}

\item{size}{point or path size. \emph{Defaults} to 2}

\item{shape}{point shape. Only relevant if single events are drawn. \emph{Default}* to 4}

\item{color}{point or path color \emph{Default} to "blue"}

\item{...}{}
}
\value{
list of ggplot geoms
}
\description{
PLots marks or colors the path ior the duration of a certain event.
}
\details{
Searches the timestamps for the closes time (more then event time)
}
