% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_limits}
\alias{geom_navr_path_limits}
\title{Adds limits based on the range of path plus and minus padding}
\usage{
geom_navr_path_limits(obj, padding)
}
\arguments{
\item{obj}{navr object}

\item{padding}{padding around the minmax of path}
}
\value{
list with xlim and ylim based on passed obj position limits and padding
}
\description{
Adds limits based on the range of path plus and minus padding
}
