% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_points}
\alias{geom_navr_path_points}
\title{Plots position points at designated times}
\usage{
geom_navr_path_points(obj, times, size = 4, shape = 18,
  color = "blue", ...)
}
\arguments{
\item{obj}{navr object}

\item{times}{Times of the positions. Takes the first position AFTER the designated time.
Times need to be on the same scale as the timestamps in \code{obj$data$timestamp}}

\item{size}{size of the point. \emph{Default} is 4}

\item{shape}{geom_point shape. \emph{Default} is 18}

\item{color}{color of the geom_point. \emph{Default} is "blue}

\item{...}{}
}
\value{
ggplot geom_point with bound data
}
\description{
Plots position points at designated times
}
