% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_rotation}
\alias{geom_navr_path_rotation}
\title{Adds rotational information to the path data}
\usage{
geom_navr_path_rotation(obj, axis = "x", downsample = 10, length = 1, ...)
}
\arguments{
\item{obj}{navr object}

\item{axis}{character designating which rotation axis to plot. Needs to follow the navr specifications.
e.g. if "z", the obj needs to have "rotation_z" column}

\item{downsample}{For very frequent data, this function would return several thousands
geoms. Therefore it is recommended to only sample each nth observation.}

\item{length}{arrow length}

\item{...}{}
}
\value{

}
\description{
Adds rotational information to the path data
}
\details{
plots arrows pointint in ditational direction. Uses \code{geom_navr_direction} function to plot each rotation
}
