% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_segments}
\alias{geom_navr_path_segments}
\title{Plots paths during certain timewindows}
\usage{
geom_navr_path_segments(obj, times, size = 2, color = "blue", ...)
}
\arguments{
\item{obj}{navr object}

\item{times}{matrix with two dimensions. Each column of the matrix marks the begining and an end
of the segment First row is the beginings and second is the ends. Path between
the beginning and end is then colored.
Times need to be on the same scale as the timestamps in \code{obj$data$timestamp}}

\item{size}{size of the path. \emph{Default} is 2}

\item{color}{color of the geom_path. \emph{Default} is "blue}

\item{...}{another optional parameters for the \code{geom_path} or \code{geom_navr_path}. See Description}
}
\value{
list of ggplot geom_paths
}
\description{
calls \link{\code{geom_navr_path}} under the hood. So any optional parameters,
such as \code{add_points}, can be also added here
}
