% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_points}
\alias{geom_navr_points}
\title{GGplot geom to add specified points to the given plot}
\usage{
geom_navr_points(ls, ...)
}
\arguments{
\item{ls}{list with XY vectors. eg. (list(start = c(0, 0), end = C(10, 5)))}

\item{...}{ggplot additional params}
}
\value{

}
\description{
GGplot geom to add specified points to the given plot
}
