% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_timeseries}
\alias{geom_navr_timeseries}
\title{Adds timeseries to the given plot}
\usage{
geom_navr_timeseries(times, values, normalise = c(), ...)
}
\arguments{
\item{times}{times on the x axis}

\item{values}{values on the y axis to plot}

\item{normalise}{if the values should be normalised to given values (min, max)}

\item{...}{}

\item{plt}{existing ggplot}
}
\description{
Adds timeseries to the given plot
}
