% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_timeseries_events}
\alias{geom_navr_timeseries_events}
\title{Draws vectical lines at times of given events}
\usage{
geom_navr_timeseries_events(event_times, ...)
}
\arguments{
\item{event_times}{times of events, need to correspond to the X axis}
}
\value{
list if geom_vline
}
\description{
Draws vectical lines at times of given events
}
