% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_position_heatmap}
\alias{geom_position_heatmap}
\title{geom to add stat_density2d position heatmap}
\usage{
geom_position_heatmap(x, y, bins = 25, ...)
}
\arguments{
\item{x}{position x}

\item{y}{position y}

\item{bins}{number of bins in each direction (n parameter in statn_density)}

\item{...}{other ggpolot arguments}
}
\description{
geom to add stat_density2d position heatmap
}
