% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{get_area_visits}
\alias{get_area_visits}
\title{Returns indices of when an area was entered for the first time}
\usage{
get_area_visits(obj, to, from = NULL, between_allowed = 0)
}
\arguments{
\item{obj}{navr object with areas.}

\item{to}{name of the area to which count entrances}

\item{from}{optional parameter defining from whih area the entrance should be}

\item{between_allowed}{number of allowed areas to be entered between the \code{from} and \code{to}.
only aplicable if \code{from} is set. Designates if  the entrance
should be counted only if it is N "stops from the \code{from} area or if any
This is primarily for the reason of defining "no" enter zones or if you have some "buffer" between.
e.g. if you want to count how many time supermarket is entered from
the "street", but the person has to pass through a "parking lot" area first, you need to
set the \code{between_allowed} to 1 or higher.}
}
\value{
indices of data where the area was entered given argument conditions or
NULL if no such visit exists
}
\description{
Returns indices of when an area was entered for the first time
}
