% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_distances.navr}
\alias{get_distances.navr}
\title{Gets distance column from the object}
\usage{
get_distances.navr(obj)
}
\arguments{
\item{obj}{navr object with calculated distances}
}
\description{
Gets distance column from the object
}
