% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_distances}
\alias{get_distances}
\title{Gets distance column from the object}
\usage{
get_distances(obj, ...)
}
\arguments{
\item{obj}{navr object with calculated distances}

\item{...}{}
}
\description{
Gets distance column from the object
}
