% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_log_timewindow}
\alias{get_log_timewindow}
\title{Gets log between determined times}
\usage{
get_log_timewindow(obj, start, end, ...)
}
\arguments{
\item{start}{start time as it appears in the log}

\item{end}{end time as it appears in the log}

\item{...}{}
}
\value{
data.frame
}
\description{
Gets log between determined times
}
