% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_n_actions}
\alias{get_n_actions}
\title{Returns how many actions of particular type were recorded}
\usage{
get_n_actions(obj, action, ...)
}
\arguments{
\item{action}{string with searched action name}

\item{...}{extra parameters}
}
\value{
integer with number of recorded actions
}
\description{
Returns how many actions of particular type were recorded
}
