% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_speeds.navr}
\alias{get_speeds.navr}
\title{Gets speed column from the object}
\usage{
get_speeds.navr(obj)
}
\arguments{
\item{obj}{navr object with calculated speeds}
}
\description{
Gets speed column from the object
}
