% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_speeds}
\alias{get_speeds}
\title{Gets speed column from the object}
\usage{
get_speeds(obj, ...)
}
\arguments{
\item{obj}{navr object with calculated speeds}

\item{...}{}
}
\description{
Gets speed column from the object
}
