% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_time_diffs.navr}
\alias{get_time_diffs.navr}
\title{Returns time_diffs column from the data if it has been caluclated}
\usage{
get_time_diffs.navr(obj)
}
\arguments{
\item{obj}{Navr Object}
}
\value{

}
\description{
Returns time_diffs column from the data if it has been caluclated
}
