% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_time_diffs}
\alias{get_time_diffs}
\title{Return time diffs column from the object}
\usage{
get_time_diffs(obj, ...)
}
\arguments{
\item{obj}{}
}
\description{
Return time diffs column from the object
}
