% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_times_since_start.navr}
\alias{get_times_since_start.navr}
\title{Returns time since start column from the navr object if it has one}
\usage{
get_times_since_start.navr(obj)
}
\arguments{
\item{obj}{Navr Object}
}
\value{
returns vector of times since start
}
\description{
Returns time since start column from the navr object if it has one
}
