% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_times_since_start}
\alias{get_times_since_start}
\title{Return time since start form the passed object}
\usage{
get_times_since_start(obj, ...)
}
\arguments{
\item{obj}{}
}
\description{
Return time since start form the passed object
}
