% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics.R
\name{get_trial_action_times}
\alias{get_trial_action_times}
\title{Returns times when certain action was performend during a particular trial}
\usage{
get_trial_action_times(obj, trialId, action, ...)
}
\arguments{
\item{action}{string with the action name}

\item{...}{}
}
\description{
Returns times when certain action was performend during a particular trial
}
