% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_distance}
\alias{get_trial_distance}
\title{REturns disntace traveled during a particular trial}
\usage{
get_trial_distance(obj, trialId, ...)
}
\arguments{
\item{...}{}
}
\value{
distance in appropriate units
}
\description{
REturns disntace traveled during a particular trial
}
