% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_goal_position}
\alias{get_trial_goal_position}
\title{Returns X and Y position of a goal location}
\usage{
get_trial_goal_position(obj, trialId, ...)
}
\arguments{
\item{...}{}
}
\value{
vector 2 of x and Y of the goal position
}
\description{
Returns X and Y position of a goal location
}
