% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_log}
\alias{get_trial_log}
\title{Gets log data.frame for a particular trial}
\usage{
get_trial_log(obj, trialId, ...)
}
\arguments{
\item{trialId}{integer with valid trialId}

\item{...}{aditional arguments}
}
\description{
Gets log data.frame for a particular trial
}
