% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_n_actions}
\alias{get_trial_n_actions}
\title{Returns how many actions of particular type were recorded during particular trial}
\usage{
get_trial_n_actions(obj, trialId, action, ...)
}
\arguments{
\item{trialId}{integer with valid trialId}

\item{action}{string with searched action name}

\item{...}{}
}
\value{
integer with number of recorded actions during particular trial
}
\description{
Returns how many actions of particular type were recorded during particular trial
}
