% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{has_areas}
\alias{has_areas}
\title{Returns if the areas have already been calculated on the object}
\usage{
has_areas(obj)
}
\arguments{
\item{obj}{valid navr object}
}
\value{
logical
}
\description{
Returns if the areas have already been calculated on the object
}
