% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{is_in_area}
\alias{is_in_area}
\title{Returns if given position is within area or not}
\usage{
is_in_area(x, y, area)
}
\arguments{
\item{x}{numeric vector of x positions. Same dimensions as y}

\item{y}{numeric vector of y positions. Same dimensions as x}

\item{area}{\code{\link{AreaObject}}}
}
\value{
logical vector of length x.
}
\description{
Returns if given position is within area or not
}
