% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-loading.R
\name{load_position_data}
\alias{load_position_data}
\title{Loads and checks position data into navr object}
\usage{
load_position_data(obj, df)
}
\arguments{
\item{df}{dataframe with data to load. Navr expects particular columns to be present, check documentation on Navr object}
}
\description{
Loads and checks position data into navr object
}
