% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-visualising.R
\name{make_circle}
\alias{make_circle}
\title{Creates data frame to be used to draw a circle}
\usage{
make_circle(center = c(0, 0), radius = 1, precision = 100)
}
\arguments{
\item{precision}{}
}
\description{
Creates data frame to be used to draw a circle
}
