% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MndUtvikling.R
\name{MndUtvikling}
\alias{MndUtvikling}
\title{R6 Class MndUtvikling}
\value{
Tabell med mnd utvikling
}
\description{
Tabell som viser manedlig endring, sammenliknet med tilsvarende periode aret for.
Periodene som vises er 3-, 6-, 9- og 12-mandersutviklingen.
}
\details{
Object MndUtviling opprettes forst, deretter lager du tabellen med method: tabellMndUtvikling
Ved kolonne == mottakere, lages tabell av mottakere, hvis regnskap, lages regnskapsutviling
}
\examples{
Opprett forst class navR::MndUtvikling$new()

Dertter navR::MndUtvikling$new()$tabellMndUtvikling()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MndUtvikling$new()}}
\item \href{#method-tabellMndUtvikling}{\code{MndUtvikling$tabellMndUtvikling()}}
\item \href{#method-giK}{\code{MndUtvikling$giK()}}
\item \href{#method-print}{\code{MndUtvikling$print()}}
\item \href{#method-clone}{\code{MndUtvikling$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MndUtvikling$new(
  df_data,
  anslag_ar = NULL,
  anslag_mnd_periode = NULL,
  REGNSKAP_ARET_FOR = NULL,
  MOTTAKERE_ARET_FOR = NULL,
  PRIS_VEKST = NULL,
  pris_gjeldende = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_data}}{data.frame med regnskap eller mottakere}

\item{\code{anslag_ar}}{Gjeldende ar for anslaget}

\item{\code{anslag_mnd_periode}}{Gjeldende maned for anslaget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabellMndUtvikling"></a>}}
\if{latex}{\out{\hypertarget{method-tabellMndUtvikling}{}}}
\subsection{Method \code{tabellMndUtvikling()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MndUtvikling$tabellMndUtvikling(tiltak_kost_ar1 = NULL, tiltak_kost_ar2 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-giK"></a>}}
\if{latex}{\out{\hypertarget{method-giK}{}}}
\subsection{Method \code{giK()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MndUtvikling$giK()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MndUtvikling$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MndUtvikling$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
