% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/navr.R
\name{navr}
\alias{navr}
\title{Responsive Toolbar htmlwidget}
\usage{
navr(selector = "body", taglist = tagList(), options = NULL, width = 0,
  height = 0)
}
\arguments{
\item{selector}{\code{string} CSS selector of html element in which to build the toolbar.}

\item{taglist}{\code{\link[htmltools]{tagList}} or \code{\link[htmltools]{HTML}}
of items for the toolbar.  Generally, these
will be \code{\link[htmltools]{tags$ul}} containing \code{\link[htmltools]{tags$li}}.}

\item{options}{\code{list} of config options for responsive-nav.js.  See
\href{https://github.com/viljamis/responsive-nav.js#usage-instructions}{Usage Instructions}
for a discussion.  Options and their default values are listed below for reference.
\itemize{
  \item animate: true // Boolean: Use CSS3 transitions, true or false
  \item transition: 284 // Integer: Speed of the transition, in milliseconds
  \item label: "Menu" // String: Label for the navigation toggle
  \item insert: "before" // String: Insert the toggle before or after the navigation
  \item customToggle: "" // Selector: Specify the ID of a custom toggle
  \item closeOnNavClick: false // Boolean: Close the navigation when one of the links are clicked
  \item openPos: "relative" // String: Position of the opened nav, relative or static
  \item navClass: "nav-collapse" // String: Default CSS class. If changed, you need to edit the CSS too!
  \item navActiveClass: "js-nav-active" // String: Class that is added to <html> element when nav is active
  \item jsClass: "js" // String: 'JS enabled' class which is added to <html> element
  \item init: function(){} // Function: Init callback
  \item open: function(){} // Function: Open callback
  \item close: function(){} // Function: Close callback
}}

\item{width}{\code{integer} width of the container in pixels.  Since this is generally expected
        to not be shown, \code{0} is the default.  However, if provided,
        \code{width} and \code{height} will be respected.}

\item{height}{\code{integer} height of the container in pixels.  Since this is generally expected
         to not be shown, \code{0} is the default.  However, if provided,
         \code{width} and \code{height} will be respected.}
}
\description{
Utility htmlwidget to provide a responsive toolbar to a htmltools::tag,
 HTML element, or another htmlwidget.  The underlying JavaScript library is
 \href{http://responsive-nav.com/}{responsive-nav.js}, which is
 lightweight and dependency-free.
}

