% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-speed-calculations.R
\name{pick_unreal_speeds}
\alias{pick_unreal_speeds}
\title{Finds indices of impossible speed values}
\usage{
pick_unreal_speeds(obj, cutoff, type = "value", ...)
}
\arguments{
\item{cutoff}{value after which we should consider speeds as invalid}

\item{type}{what type of cutoff. Possibilities value (above value is unreal),
percent (move than certain percent increase), std (number of stds away from mean speed),
quantile(percent of highest values)}

\item{speeds}{speed vector of doubles}
}
\value{
indices of speeds considered unreal
}
\description{
Finds indices of impossible speed values
}
