% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_background}
\alias{plot_add_background}
\title{Adds image to the background of the plot}
\usage{
plot_add_background(plt, image_path, xlim = NULL, ylim = NULL)
}
\arguments{
\item{plt}{existing plot}

\item{image_path}{path the png image}

\item{xlim}{places the image on the background}

\item{ylim}{}
}
\description{
Adds image to the background of the plot
}
