% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_image}
\alias{plot_add_image}
\title{Adds image to the background of the plot}
\usage{
plot_add_image(plt, image_path, xlim, ylim)
}
\arguments{
\item{plt}{existing plot}

\item{image_path}{path the png image}

\item{ylim}{}
}
\description{
Adds image to the background of the plot
}
