% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_limits}
\alias{plot_add_limits}
\title{Checks if object has map limits variable and adds plot limits if so}
\usage{
plot_add_limits(plt, obj)
}
\arguments{
\item{plt}{existing plot}

\item{obj}{navr object}
}
\description{
Checks if object has map limits variable and adds plot limits if so
}
