% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_add_path}
\alias{plot_add_path}
\title{Adds path to the plot data}
\usage{
plot_add_path(plt, obj, ...)
}
\arguments{
\item{plt}{}

\item{obj}{navr object to which}
}
\description{
Adds path to the plot data
}
