% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_points}
\alias{plot_add_points}
\title{Adds specified points to the given plot}
\usage{
plot_add_points(plt, ls, size = 4, color = "blue")
}
\arguments{
\item{plt}{previous ggplot}

\item{ls}{list with XY vectors. eg. (list(start = c(0, 0), end = C(10, 5)))}

\item{size}{size of the dot}

\item{color}{dot color}
}
\value{
modified plot
}
\description{
Adds specified points to the given plot
}
