% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_add_position_heatmap}
\alias{plot_add_position_heatmap}
\title{PLots heatmap using stat_density2d}
\usage{
plot_add_position_heatmap(plt, obj, bins, ...)
}
\arguments{
\item{plt}{plot to which to add}

\item{obj}{navr object which to plot}

\item{bins}{number of bins in each direction (n parameter in statn_density)}
}
\description{
PLots heatmap using stat_density2d
}
