% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_shape}
\alias{plot_add_shape}
\title{Adds shape to the background of the plot}
\usage{
plot_add_shape(plt, x, y, ...)
}
\arguments{
\item{plt}{existing plot}

\item{x}{vector with pologyon X coordinates}

\item{y}{vector with pologyon Y coordinates}

\item{...}{polygon aesthetics, such as fill, color, alpha etc.}
}
\description{
Adds shape to the background of the plot
}
