% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_position_heatmap.navr}
\alias{plot_position_heatmap.navr}
\title{Plots position heatmap with given granularity}
\usage{
\method{plot_position_heatmap}{navr}(obj, bins = 25, ...)
}
\arguments{
\item{obj}{object}

\item{bins}{number of bins to segment the area}

\item{...}{optional ggplot parameters}
}
\value{
ggplot
}
\description{
Plots position heatmap with given granularity
}
