% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_position_heatmap}
\alias{plot_position_heatmap}
\title{Plots position heatmap with given granularity}
\usage{
plot_position_heatmap(obj, bins, ...)
}
\arguments{
\item{obj}{object}

\item{bins}{number of bins to segment the area}

\item{...}{optional ggplot parameters}
}
\value{

}
\description{
Plots position heatmap with given granularity
}
