% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_speed.navr}
\alias{plot_speed.navr}
\title{PLots speed values in time}
\usage{
\method{plot_speed}{navr}(obj, normalise = c(), ...)
}
\arguments{
\item{normalise}{if the values should be normalised to certain values - needs vector of length 2 c(0,1)}

\item{...}{}
}
\value{
ggplot
}
\description{
PLots speed values in time
}
