% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-plotting.R
\name{plot_trial_path}
\alias{plot_trial_path}
\title{Generic function}
\usage{
plot_trial_path(obj, trialId, ...)
}
\arguments{
\item{trialId}{vector of trial ids}

\item{...}{Optional parameters}
}
\value{
ggplot object
}
\description{
Generic function
}
