% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-loading.R
\name{prepare_column_names}
\alias{prepare_column_names}
\title{Tries to rename columns so they correspond to proper naming conventions}
\usage{
prepare_column_names(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
modified dataframe
}
\description{
changes "." to "_" to correspond with python conventions,
makes everything lowecase, renames Time column if present to "timestamp"
}
