% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{prepare_navr}
\alias{prepare_navr}
\title{Adds all necessary columns to the navr object}
\usage{
prepare_navr(obj)
}
\arguments{
\item{obj}{object of type navr}
}
\value{
returns navr object with modified columns
}
\description{
Adds all necessary columns to the navr object
}
