% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{radian_to_angle}
\alias{radian_to_angle}
\title{Converts Pi radian angle to degrees}
\usage{
radian_to_angle(radian)
}
\arguments{
\item{radian}{}
}
\value{

}
\description{
Converts Pi radian angle to degrees
}
