% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegnskapTabell.R
\name{Regnskapstabell}
\alias{Regnskapstabell}
\alias{RegnskapTabell}
\title{Regnskapstabell}
\value{
Tabell med årlige regnskapstall, og per mnd.
}
\description{
Regnskapstabell

Regnskapstabell
}
\examples{
Dette er en test. Kommer dette i dokumentasjonen?
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RegnskapTabell$new()}}
\item \href{#method-lagRegnskapTabell}{\code{RegnskapTabell$lagRegnskapTabell()}}
\item \href{#method-print}{\code{RegnskapTabell$print()}}
\item \href{#method-clone}{\code{RegnskapTabell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$new(
  dfRegnskap,
  g_gjeldende,
  anslag_ar,
  anslag_mnd_periode,
  post = "post"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dfRegnskap}}{data.frame med regnskapstall per mnd Tre rader: Nominelle regnskapstall, dato "YYYY-MM-DD" og mnd prisjusterer.}

\item{\code{g_gjeldende}}{Forutsatt prisjusterer.}

\item{\code{anslag_ar}}{Gjeldende år for anslaget}

\item{\code{anslag_ar}}{Gjeldende måned for anslaget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagRegnskapTabell"></a>}}
\if{latex}{\out{\hypertarget{method-lagRegnskapTabell}{}}}
\subsection{Method \code{lagRegnskapTabell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$lagRegnskapTabell()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
