% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegnskapTabell.R
\name{RegnskapTabell}
\alias{RegnskapTabell}
\title{R6 Class Regnskapstabell}
\value{
Tabell med arlige summerte regnskapstall
}
\description{
Tabellen viser nominelle og prisjustert utgifter. Per ar og per mnd i innevaerende ar.
}
\details{
Tabellen ma "lages" med metode "lagTabell".
}
\examples{
Opprett forst class regnskapstabell:RegnskapTabell$new(
   dfRegnskap = navR::regnskap,
   g_gjeldende = 104817,
   anslag_ar = 2021,
   anslag_mnd_periode = 2,
   post = "post70"
   )

Dertter lag tabell med metode:
RegnskapTabell$lagRegnskapTabell()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RegnskapTabell$new()}}
\item \href{#method-lagRegnskapTabell}{\code{RegnskapTabell$lagRegnskapTabell()}}
\item \href{#method-print}{\code{RegnskapTabell$print()}}
\item \href{#method-clone}{\code{RegnskapTabell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$new(
  dfRegnskap,
  g_gjeldende,
  anslag_ar,
  anslag_mnd_periode,
  post = "post70"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dfRegnskap}}{data.frame med regnskapstall per mnd. input data skal ha tre rader  dato ("YYYY-MM-DD"), nominelle regnskapstall per mnd og prisjusterer per mnd.}

\item{\code{g_gjeldende}}{Forutsatt prisjusterer.}

\item{\code{anslag_ar}}{Gjeldende ar for anslaget}

\item{\code{anslag_mnd_periode}}{Gjeldende maned for anslaget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lagRegnskapTabell"></a>}}
\if{latex}{\out{\hypertarget{method-lagRegnskapTabell}{}}}
\subsection{Method \code{lagRegnskapTabell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$lagRegnskapTabell()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegnskapTabell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
