% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{remove_unreal_speeds}
\alias{remove_unreal_speeds}
\title{Shorthand for picking speeds and then recalculating object}
\usage{
remove_unreal_speeds(obj, ...)
}
\arguments{
\item{...}{}
}
\description{
Shorthand for picking speeds and then recalculating object
}
