% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-loading.R
\name{rename_column}
\alias{rename_column}
\title{Helper to rename column}
\usage{
rename_column(df, old_column, new_column)
}
\arguments{
\item{df}{input datafgrames}

\item{old_column}{old column name}

\item{new_column}{new column name}
}
\value{
modified dataframe
}
\description{
Helper to rename column
}
