% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/navr.R
\name{renderNavr}
\alias{renderNavr}
\title{Widget render function for use in Shiny}
\usage{
renderNavr(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

