% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-helpers.R
\name{replace_na}
\alias{replace_na}
\title{Replaces na values with given type}
\usage{
replace_na(vec, replacement = "last.known", ...)
}
\arguments{
\item{...}{}

\item{replacement:}{"last.known", "number"}
}
\description{
Replaces na values with given type
}
