% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-helpersl.R
\name{rolling_sum}
\alias{rolling_sum}
\title{Calculates rolling sum of a vector across N points}
\usage{
rolling_sum(vec, n_points)
}
\arguments{
\item{vec}{numeric vector across which we calculate}

\item{n_points}{}
}
\value{
rolled vector sum or NULL if the n_points is larger than length of the vector
}
\description{
Calculates rolling sum of a vector across N points
}
