% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-onsets.R
\name{search_onsets}
\alias{search_onsets}
\title{Searches for times when the movement onset started}
\usage{
search_onsets(obj, speed_threshold, min_duration, ...)
}
\arguments{
\item{...}{}
}
\value{
list with times (time since start) of onset and durations of movement
}
\description{
Searches for times when the movement onset started
}
