% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-onsets.R
\name{search_stops.navr}
\alias{search_stops.navr}
\title{Searches for movement stops and returns time since start for each event}
\usage{
\method{search_stops}{navr}(obj, speed_threshold, min_duration = 0)
}
\arguments{
\item{obj}{navr object with calculated speeds}

\item{speed_threshold}{what is the speed considered to be the moving speed}

\item{min_duration}{in secouds how long should the person be still}
}
\value{
list with times (time since start) and durations of stillness
}
\description{
Searches for movement stops and returns time since start for each event
}
