% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-onsets.R
\name{search_stops}
\alias{search_stops}
\title{Searches for times when there is no movement}
\usage{
search_stops(obj, speed_threshold, min_duration, ...)
}
\arguments{
\item{...}{}
}
\value{
list with times (time since start) and durations of stillness
}
\description{
Searches for times when there is no movement
}
