% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-onsets.R
\name{search_stops_speeds_times}
\alias{search_stops_speeds_times}
\title{Searches for times of stillness from passed speeds and time diffs}
\usage{
search_stops_speeds_times(speeds, time_diffs, speed_threshold,
  min_duration)
}
\arguments{
\item{speeds}{vector of speeds}

\item{time_diffs}{vactor of time diffs}

\item{speed_threshold}{what is considered to eb moving speed}

\item{min_duration}{minímum stillness duration to be considered stop}
}
\description{
Searches for times of stillness from passed speeds and time diffs
}
