% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_positions.navr}
\alias{smooth_positions.navr}
\title{Smooth positions in a navr object}
\usage{
\method{smooth_positions}{navr}(obj, type, ....)
}
\arguments{
\item{obj}{navr object}

\item{type}{string: median, approx, spline}

\item{....}{}
}
\description{
Smooth positions in a navr object
}
