% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_positions}
\alias{smooth_positions}
\alias{smooth_positions.navr}
\title{Smooths positions of a given boject}
\usage{
smooth_positions(obj, type, ...)

\method{smooth_positions}{navr}(obj, type, ....)
}
\arguments{
\item{obj}{valid object to be smoothed}

\item{type}{string: median, approx, spline}

\item{....}{optional parameters for smoothing}
}
\description{
Smooths positions of a given boject
}
\section{Methods (by class)}{
\itemize{
\item \code{navr}: position smoothing in the navr object
}}

