% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_positions_df}
\alias{smooth_positions_df}
\title{Smooths positions in a data frame}
\usage{
smooth_positions_df(df, col_x, col_y, type, ...)
}
\arguments{
\item{df}{Data frame to modify}

\item{col_x}{name of position of the df column with X variable}

\item{col_y}{name of position of the df column with Y variable}

\item{type}{median, approx, spline}
}
\description{
Smooths positions in a data frame
}
