% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_speed.navr}
\alias{smooth_speed.navr}
\title{Smooths speed vector in the navr object}
\usage{
\method{smooth_speed}{navr}(obj, type, ...)
}
\arguments{
\item{obj}{navr object}

\item{type}{type of smoothing}

\item{...}{optional parameters for given smoothing, see `smooth_vector` function}
}
\value{
modified navr object
}
\description{
Smooths speed vector in the navr object
}
