% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_speed}
\alias{smooth_speed}
\title{Smooths speed}
\usage{
smooth_speed(obj, type, ...)
}
\arguments{
\item{obj}{object to perfrom speed smoothing on}

\item{type}{what type of smoothing to perform}

\item{...}{}
}
\description{
Smooths speed
}
