% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_vector}
\alias{smooth_vector}
\title{Smooths positions of given vector}
\usage{
smooth_vector(x, type, ...)
}
\arguments{
\item{x}{vector to smooth}

\item{type}{median, approx, spline}

\item{...}{optional parameters for the smoothing function (see below)}

\item{df}{Data frame to modify}
}
\value{
smoothed vector
}
\description{
Smooths positions of given vector
}
\details{
median takes parameters (points)
}
