% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{theme_invisible}
\alias{theme_invisible}
\title{Completely empty theme for clear plotting}
\usage{
theme_invisible(base_size = 12)
}
\arguments{
\item{base_size}{}
}
\description{
Completely empty theme for clear plotting
}
