% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{vector_from_angle}
\alias{vector_from_angle}
\title{Creates x y coordinate of point position given angle X}
\usage{
vector_from_angle(angle)
}
\arguments{
\item{angle}{numeric angle in degrees (0-360)}
}
\value{
vector
}
\description{
Creates x y coordinate of point position given angle X
}
