% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-areas.R
\name{add_areas}
\alias{add_areas}
\title{Adds information about what area does the current position belong to}
\usage{
add_areas(obj, areas, ...)
}
\arguments{
\item{areas}{list of \coode{\link{AreaObject}}. Needs to be a list of length(areas), even if you are passing
a single area, you should pass it as a list(your_area)}

\item{...}{}
}
\description{
The column contains information about in which area the
}
