% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{angle_from_positions}
\alias{angle_from_positions}
\title{Calculates angle}
\usage{
angle_from_positions(pos_from, pos_to, zero_vec = c(0, 1))
}
\arguments{
\item{zero_vec}{defines which axis should correspond to 0 degrees. defaults to c(0,1) (Y axis)}
}
\description{
Calculates angle
}
