% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{animate_path}
\alias{animate_path}
\title{Animates plot built with plot_path function}
\usage{
animate_path(plt, ...)
}
\arguments{
\item{plt}{Needs to have timestamp x value}

\item{...}{}
}
\description{
The timing of the plot animation is built arount the plot$data$timestamp field,
which is added in the `plot_path` function. If you didn't create your plot in this way, you can either
add the field to the ggplot list data, or you can animate the path in a different way
}
\details{
Calls `transition_reveal` from the gganimate package to plot the path, but
}
