% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-angle-calculations.R
\name{calculate_angle_differences}
\alias{calculate_angle_differences}
\title{Creates a series of angle_diffs conversted to degrees of angukar difference where it calculates angle difference between rows}
\usage{
calculate_angle_differences(rotations)
}
\arguments{
\item{rotations}{vector of angles in 360}
}
\description{
Creates a series of angle_diffs conversted to degrees of angukar difference where it calculates angle difference between rows
}
