% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-helpers.R
\name{euclid_distance}
\alias{euclid_distance}
\title{Euclidian distance between two points}
\usage{
euclid_distance(point1, point2)
}
\arguments{
\item{point1}{numeric position (usually length 2 or 3)}

\item{point2}{numeric position (usually length 2 or 3)}
}
\value{
numeric or NA if not valid
}
\description{
Calculates euclidian distance between two points.
}
\examples{
print(euclid_distance(c(0,0), c(1,1)))
}
