% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{filter_times}
\alias{filter_times}
\title{Filters navr object to only include times in between}
\usage{
filter_times(obj, times, zero_based, ...)
}
\arguments{
\item{times}{a vector of length 2, start time and end time in seconds}

\item{zero_based}{signifying if you passed seconds from start or reaů time. defualts to False}

\item{...}{aditional arguments}
}
\description{
Filters navr object to only include times in between
}
