% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_obj_timeseries}
\alias{geom_navr_obj_timeseries}
\title{Adds timeseries to the given plot from navr object based on given column}
\usage{
geom_navr_obj_timeseries(
  obj,
  colname,
  scaling = "none",
  constraints = NULL,
  ...
)
}
\arguments{
\item{obj}{NavrObject}

\item{colname}{name of the column to be plotted}

\item{scaling}{type of scaling to implemnent. "none", "std", "minmax"}

\item{constraints}{if the values should be scaled to given values}

\item{...}{}
}
\description{
Adds timeseries to the given plot from navr object based on given column
}
\details{
wrapper around \code{geom_navr_timeseries} for easier acces to the object
}
