% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path}
\alias{geom_navr_path}
\title{Geom of plotting navr path}
\usage{
geom_navr_path(obj, add_points = F, ...)
}
\arguments{
\item{obj}{Navr object}

\item{add_points}{if points should be explicitely noted on the path}
}
\description{
Geom of plotting navr path
}
