% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_path_events}
\alias{geom_navr_path_events}
\title{Plots positions of given events on a path}
\usage{
geom_navr_path_events(obj, event_times, size = 2, shape = 18,
  color = "blue", ...)
}
\arguments{
\item{obj}{navr objects}

\item{event_times}{numeric vector of event times. Times need to be on the same scale as the timestamps in `obj$data$timestamp`}

\item{size}{point size **default** 2}

\item{shape}{point shape **default** is 4}

\item{color}{point color **default** is "blue"}

\item{...}{}
}
\description{

}
\details{
Searches the timestamps for the closes time (more then event time)
}
