% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{geom_navr_timeseries}
\alias{geom_navr_timeseries}
\title{Adds timeseries to the given plot}
\usage{
geom_navr_timeseries(times, values, scaling = "none", constraints = NULL, ...)
}
\arguments{
\item{times}{times on the x axis}

\item{values}{values on the y axis to plot}

\item{scaling}{type of scaling to implemnent. "none", "std", "minmax"}

\item{constraints}{if the values should be scaled to given values}

\item{...}{}
}
\value{

}
\description{
Adds timeseries to the given plot
}
