% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_duration}
\alias{get_trial_duration}
\title{Returns duration of the trial in s}
\usage{
get_trial_duration(obj, trialId, ...)
}
\arguments{
\item{...}{}
}
\value{
time in s
}
\description{
Returns duration of the trial in s
}
