% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics.R
\name{get_trial_start_position}
\alias{get_trial_start_position}
\title{Returns X and Y of a starting position}
\usage{
get_trial_start_position(obj, trialId, ...)
}
\arguments{
\item{...}{}
}
\value{
vector 2 of x and Y of the start position
}
\description{
Returns X and Y of a starting position
}
