% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-generics-getting.R
\name{get_trial_times}
\alias{get_trial_times}
\title{Returns list with trial times}
\usage{
get_trial_times(obj, trialId, ...)
}
\arguments{
\item{...}{}
}
\value{
list with start and end field
}
\description{
Returns list with trial times
}
