% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-plot-elements.R
\name{plot_add_path}
\alias{plot_add_path}
\title{Adds path to the plot data}
\usage{
plot_add_path(plt, x, y)
}
\arguments{
\item{plt}{}

\item{x}{vector of positions}

\item{y}{vector of y positions}
}
\description{
Adds path to the plot data
}
