% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_path.navr}
\alias{plot_path.navr}
\title{plots the _x and _X coordinates}
\usage{
\method{plot_path}{navr}(obj, add_points = FALSE, ...)
}
\arguments{
\item{obj}{valid navr object}

\item{add_points}{if points should be explicitely noted on the path}

\item{...}{optional ggplot parameters for the path geom}
}
\value{
ggplot
}
\description{
plots the _x and _X coordinates
}
