% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_speed}
\alias{plot_speed}
\alias{plot_speed.navr}
\title{PLots speed values in time}
\usage{
plot_speed(obj, scaling = "none", constraints = NULL, ...)

\method{plot_speed}{navr}(obj, scaling = "none", constraints = NULL, ...)
}
\arguments{
\item{obj}{valid object}

\item{scaling}{type of scaling to implement, possibilities are "none", "std", "minmax".
\emph{Default} is "none"}

\item{constraints}{if the values should be scaled to certain values - needs numeric(2)}

\item{...}{additional parameters for the \code{\link{ggplot::geom_line}}}
}
\value{
ggplot plot
}
\description{
PLots speed values in time

PLots speed values in time
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_speed(navr)}: Plot speed values for the nvar object

}}
