% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-visualising.R
\name{plot_speed}
\alias{plot_speed}
\title{PLots speed values in time}
\usage{
plot_speed(obj, scaling = "none", scale = c(), ...)
}
\arguments{
\item{scaling}{type of scaling to implement, possibilities are "std", "minmax"}

\item{scale}{if the values should be scaled to certain values - needs vector of length 2 c(0,1)}
}
\value{
ggplot
}
\description{
PLots speed values in time
}
