% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-data-modifications.R
\name{remove_unreal_speeds.navr}
\alias{remove_unreal_speeds.navr}
\title{Inserts NA values to speed and distance}
\usage{
\method{remove_unreal_speeds}{navr}(obj, indices, total_recalculate = T)
}
\arguments{
\item{indices}{indices of speeds to clean out}

\item{total_recalculate}{if true, recalculates total_distance column to reflect removed distances}
}
\value{
navr object with NA values in appropriate places
}
\description{
Inserts NA values to speed and distance
}
