% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-onsets.R
\name{search_onsets.navr}
\alias{search_onsets.navr}
\title{Title}
\usage{
\method{search_onsets}{navr}(obj, speed_threshold, min_duration = 0,
  return_duration = F, still_speed_threshold = speed_threshold,
  still_duration = 0)
}
\arguments{
\item{obj}{navr object with calculated speeds}

\item{still_duration}{}
}
\description{
Title
}
