% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_positions}
\alias{smooth_positions}
\title{Smooths positions of a given boject}
\usage{
smooth_positions(obj, ...)
}
\arguments{
\item{obj}{}
}
\description{
Smooths positions of a given boject
}
