% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navr-smoothing.R
\name{smooth_vector}
\alias{smooth_vector}
\title{Smooths positions of given vector}
\usage{
smooth_vector(x, type, ...)
}
\arguments{
\item{x}{numeric vector to smooth}

\item{type}{median, approx, spline}

\item{...}{optional parameters for the smoothing function (see below)}
}
\value{
smoothed numeric vector
}
\description{
Smooths positions of given vector
}
\details{
\emph{median} takes parameters points (default is 11). See \code{\link{runmed}}
\emph{spline} takes parameters and spar, nknots. See \code{\link{smooth.spline}}
}
